requirements.txt
daemon
feedparser
python-twitter
sphinx-me
PyYAML
